# api/openapi.json
{
  "components": {
    "schemas": {
...
      "Corridor": {
        "properties": {
          "congestion_index": {
            "title": "Congestion Index",
            "type": "number"
          },
          "free_flow_s": {
            "title": "Free Flow S",
            "type": "number"
          },
          "from_camera": {
            "title": "From Camera",
            "type": "string"
          },
          "median_travel_s": {
            "title": "Median Travel S",
            "type": "number"
          },
          "n_trips": {
            "title": "N Trips",
            "type": "integer"
          },
          "p90_travel_s": {
            "title": "P90 Travel S",
            "type": "number"
          },
          "to_camera": {
            "title": "To Camera",
            "type": "string"
          }
        },
        "required": [
          "from_camera",
          "to_camera",
          "n_trips",
          "median_travel_s",
          "p90_travel_s",
          "free_flow_s",
          "congestion_index"
        ],
        "title": "Corridor",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VolumeBucket": {
        "properties": {
          "bucket_start": {
            "title": "Bucket Start",
            "type": "string"
          },
          "camera_id": {
            "title": "Camera Id",
            "type": "string"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "camera_id",
          "bucket_start",
          "count"
        ],
        "title": "VolumeBucket",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SUTRA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/analytics/corridors": {
      "get": {
        "operationId": "corridors_api_analytics_corridors_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Corridor"
                  },
                  "title": "Response Corridors Api Analytics Corridors Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Corridors",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/api/analytics/volumes": {
      "get": {
        "operationId": "volumes_api_analytics_volumes_get",
        "parameters": [
          {
            "in": "query",
            "name": "bucket_minutes",
            "required": false,
            "schema": {
              "default": 60,
              "title": "Bucket Minutes",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "camera_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Camera Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VolumeBucket"
                  },
                  "title": "Response Volumes Api Analytics Volumes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Volumes",
        "tags": [
          "analytics"
        ]
      }
    },
...
  }
}